# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Personal Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DAILY PLANS ──────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
        - name: to
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CHORES ───────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Habits ───────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, userId, title, status, progress, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, paused]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        category:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GoalInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        targetDate:
          type: string
        category:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        targetDate:
          type: string
        status:
          type: string
          enum: [active, completed, paused]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        category:
          type: string

    # ── Daily Plans ───────────────────────────────────────────────────────
...
    Event:
      type: object
      required: [id, userId, title, startDate, allDay, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        startTime:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        allDay:
          type: boolean
        location:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        attendees:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [title, startDate]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
        startTime:
          type: string
        endDate:
          type: string
        endTime:
          type: string
        allDay:
          type: boolean
        location:
          type: string
        color:
          type: string
        attendees:
          type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
        startTime:
          type: string
        endDate:
          type: string
        endTime:
          type: string
        allDay:
          type: boolean
        location:
          type: string
        color:
          type: string
        attendees:
          type: string

    # ── Chores ────────────────────────────────────────────────────────────
...