# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  targetDate?: string;
  category?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
}